pyproject.toml
[tool.poetry]
name = "db-course-work"
version = "0.1.0"
description = ""
authors = ["vvvadim <v.vladimirov@it-park.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.31"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
